Remove spaces, rename some functions to align it more with the functionality
