+++ cmd/es-rollover/app/init/flags.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package initialize
+package init
 
 import (
 	"flag"
