+++ cmd/es-rollover/app/rollover/action.go
@@ -32,14 +32,14 @@ type Action struct {
 func (a *Action) Do() error {
 	rolloverIndices := app.RolloverIndices(a.Config.Archive, a.Config.IndexPrefix)
 	for _, indexName := range rolloverIndices {
-		if err := a.action(indexName); err != nil {
+		if err := a.rollover(indexName); err != nil {
 			return err
 		}
 	}
 	return nil
 }
 
-func (a *Action) action(indexSet app.IndexOption) error {
+func (a *Action) rollover(indexSet app.IndexOption) error {
 	conditionsMap := map[string]interface{}{}
 	if len(a.Conditions) > 0 {
 		err := json.Unmarshal([]byte(a.Config.Conditions), &conditionsMap)
